[Calculator Form Application - Exponent Precedence and Double]
Add exponent operation and precedence to the expression builder.
Changed the Number class to now use Double instead of int for floating point operations.
Removed the Calculator class as it had no use with current functionality.
Calculator input was processed to correctly clear the text field depending on input and last operation performed.

+++ CalculatorForm/CalculatorForm/Parsing/Number.cs
@@ -2,14 +2,14 @@
 {
     class Number : IExpression
     {
-        private int value;
+        private double value;
         
-        public Number(int value)
+        public Number(double value)
         {
             this.value = value;
         }
 
-        public int evaluate()
+        public double evaluate()
         {
             return value;
         }
